refactor : 로그인 버튼 제거 및 컴포넌트 분리
